+++ test/integration/parameter.cpp
@@ -2,8 +2,8 @@
 // Distributed under the Boost Software License, Version 1.0.
 // (See accompanying file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 
-#include "../getenv.hpp"
-#include "../macros.hpp"
+#include "utils/getenv.hpp"
+#include "utils/macros.hpp"
 
 #include <exception>
 #include <iostream>
