--- test/pq/CMakeLists.txt
@@ -1,18 +1,0 @@
-cmake_minimum_required(VERSION 3.15)
-
-file(GLOB testsources *.cpp)
-foreach(testsourcefile ${testsources})
-  get_filename_component(exename taopq-test-${testsourcefile} NAME_WE)
-  add_executable(${exename} ${testsourcefile})
-  target_link_libraries(${exename} PRIVATE taocpp::taopq)
-  target_compile_features(${exename} PRIVATE cxx_std_20)
-  if(MSVC)
-    target_compile_options(${exename} PRIVATE /W4 /WX /utf-8)
-  else()
-    target_compile_options(${exename} PRIVATE -pedantic -Wall -Wextra -Wshadow -Werror)
-  endif()
-  if(WIN32)
-    target_link_libraries(${exename} PRIVATE wsock32 ws2_32)
-  endif()
-  add_test(NAME ${exename} WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/../../.. COMMAND ${CMAKE_CURRENT_BINARY_DIR}/${exename})
-endforeach(testsourcefile)
