+++ test/unit/resize_uninitialized.cpp
@@ -2,7 +2,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // (See accompanying file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 
-#include "../macros.hpp"
+#include "utils/macros.hpp"
 
 #include <cstddef>
 #include <exception>
