+++ test/unit/result_type.cpp
